requirements.txt
streamlit>=1.28.0
pandas>=1.5.0
numpy>=1.24.0
scikit-learn>=1.3.0
matplotlib>=3.7.0
seaborn>=0.12.0
plotly>=5.15.0
supabase>=2.0.0
python-dotenv>=1.0.0
pathlib2>=2.3.7
pickle-mixin>=1.0.2